// src/app/types/contact.ts
export type SocialLink = {
    socialLink: string
  }
  
  export type ContactInfo = {
    heading: string
    addressLine1: string
    addressLine2: string
    phone: string
    hours: string
    email: string
    note: string
  }
  
  export type ContactFormData = {
    name: string
    email: string
    message: string
  }